+++ test/wrappers/process-contract-params.ts
@@ -51,7 +51,7 @@ describe("Process contract parameter wrapper", () => {
         const params2 = ProcessContractParameters.fromParams({
             mode: 9,
             envelopeType: 2,
-            censusOrigin: 4,
+            censusOrigin: 1,
             // tokenAddress
             metadata: "65",
             censusRoot: "87",
@@ -71,7 +71,7 @@ describe("Process contract parameter wrapper", () => {
 
         expect(params2[0][0]).to.eq(9)
         expect(params2[0][1]).to.eq(2)
-        expect(params2[0][2]).to.eq(4)
+        expect(params2[0][2]).to.eq(1)
         expect(params2[1]).to.eq("0x0000000000000000000000000000000000000000")
         expect(params2[2][0]).to.eq("65")
         expect(params2[2][1]).to.eq("87")